requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-log==0.4.0
colorama==0.4.6
crewai==0.5.0
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
docutils==0.21.2
dotty-dict==1.3.1
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gotrue==1.3.1
greenlet==3.1.1
gspread==5.11.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
idna==3.10
importlib_metadata==8.6.1
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
lxml==5.3.2
marshmallow==3.26.1
more-itertools==10.6.0
multidict==6.4.3
mypy-extensions==1.0.0
nh3==0.2.21
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.73.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pillow==11.2.1
pkginfo==1.12.1.2
playwright==1.51.0
postgrest==0.10.7
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2025.2
PyYAML==6.0.1
readme_renderer==44.0
realtime==1.0.6
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
selenium==4.31.0
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tenacity==8.5.0
tiktoken==0.5.2
tomlkit==0.13.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
twine==3.8.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.19.0
zipp==3.21.0
